use TPL_HIGH_LEVEL for spinlocks to make threading interrupt-safe
Since we still need to call functions which are only available in TPL_NOTIFY,
we check if we were called using max TPL_CALLBACK and lower the TPL
just for these functions.

Should an interrupt happen during that time and the handler calls
threading code we just assert because interrupt handlers are not
allowed to do that anyway.